requirements.txt
python-digitalocean==1.17.0
requests==2.28.1
dynaconf==3.1.11